+++ mostlyai/sdk/_data/constraints/types/base.py
